use lstat when removing the old socket
